Added dialogue choice test designed for exercising issues with cursor positioning.
Fixed a bug where a page of only choices and no dialogue text was included, it would error out.

Fixed cursor positioning so that a page of only choices will result in the choice and cursor being on the first line.

Added some code to override the skip the opening dialogue and run the choices test.